feat: add amplify/redux user slice
This should probably be reworked at some stage to integrate with
the Amplify APIs directly. However this seems to work and forwards
the current user to the user reducer; where we can then make the
rest of the app.

This commit also deals with the slightly horrendous setup of
redux with next.