fix: sort out 'linting' issues in production
Next.js provides their own integration which is apparently opinionated
enough to fail builds in production but not in dev. The type checking
is also not as strict in dev as prod :/

We should apply a build guard before PR merge - until then this _has_
been checked with a local production build.