perf: improve test performance by 74%
As of this commit, this decreases the ~ 5.9 s test time down
to 2.7 s. A 74% improvement! The issue, is that jest has to
resolve the _entire_ collection of icons for each test file.
As the UserPopoverMenu imports a few of these, we can make
massive gains by simply mocking the component implementations!