feat: trial useDependencies hook for DI
We needed to pass injected hooks increasing levels
down the tree, which was getting bloated, repeated, and a
lot of boilerplate on a case by case basis. This allows hooks,
and in theory anything, to be globally registered and injected
into the relevant component. The global dependencies are mocked
away and the runtime values can be freely passed using props of
DependencyProvider.